Refactored output writers to pass tests.
