+++ test/runtests.jl
@@ -219,39 +219,6 @@ float_types = [Float32, Float64]
         end
     end
 
-    @testset "Boundary conditions" begin
-        include("test_boundary_conditions.jl")
-
-        Nx, Ny, Nz = 3, 4, 5 # for simple test
-        funbc(args...) = π
-
-        for fld in (:u, :v, :T, :S)
-            for bctype in (Gradient, Flux)
-                for bc in (0.6, rand(Nx, Ny), funbc)
-                    @test test_z_boundary_condition_simple(fld, bctype, bc, Nx, Ny, Nz)
-                end
-            end
-            @test test_diffusion_simple(fld)
-            @test test_diffusion_budget(fld)
-            @test test_diffusion_cosine(fld)
-            @test test_flux_budget(fld)
-        end
-    end
-
-    @testset "Forcing" begin
-        add_one(args...) = 1.0
-        function test_forcing(fld)
-            kwarg = Dict(Symbol(:F, fld)=>add_one)
-            forcing = Forcing(; kwarg...)
-            f = getfield(forcing, fld)
-            f() == 1.0
-        end
-
-        for fld in (:u, :v, :w, :T, :S)
-            @test test_forcing(fld)
-        end
-    end
-
     @testset "Time stepping" begin
         println("  Testing time stepping...")
         include("test_time_stepping.jl")
@@ -267,26 +234,67 @@ float_types = [Float32, Float64]
         end
     end
 
+    @testset "Boundary conditions" begin
+        println("  Testing boundary conditions...")
+        include("test_boundary_conditions.jl")
+
+        Nx, Ny, Nz = 3, 4, 5 # for simple test
+        funbc(args...) = π
+
+        for fld in (:u, :v, :T, :S)
+            for bctype in (Gradient, Flux)
+                for bc in (0.6, rand(Nx, Ny), funbc)
+                    @test test_z_boundary_condition_simple(fld, bctype, bc, Nx, Ny, Nz)
+                end
+            end
+
+            @test test_diffusion_simple(fld)
+            @test test_diffusion_budget(fld)
+            @test test_diffusion_cosine(fld)
+            @test test_flux_budget(fld)
+        end
+    end
+
+    @testset "Forcing" begin
+        println("  Testing forcings...")
+        add_one(args...) = 1.0
+        function test_forcing(fld)
+            kwarg = Dict(Symbol(:F, fld)=>add_one)
+            forcing = Forcing(; kwarg...)
+            f = getfield(forcing, fld)
+            f() == 1.0
+        end
+
+        for fld in (:u, :v, :w, :T, :S)
+            @test test_forcing(fld)
+        end
+    end
+
+
+    @testset "Output writers" begin
+        println("  Testing output writers...")
         include("test_output_writers.jl")
 
         @testset "Checkpointing" begin
+            println("    Testing checkpointing...")
             run_thermal_bubble_checkpointer_tests()
         end
 
         @testset "NetCDF" begin
+            println("    Testing NetCDF output writer...")
             run_thermal_bubble_netcdf_tests()
         end
     end
 
-    @testset "Golden master tests" begin
-        include("test_golden_master.jl")
-
-        @testset "Thermal bubble" begin
-            run_thermal_bubble_golden_master_tests()
-        end
-
-        @testset "Deep convection" begin
-            run_deep_convection_golden_master_tests()
-        end
-    end
+    # @testset "Golden master tests" begin
+    #     include("test_golden_master.jl")
+    #
+    #     @testset "Thermal bubble" begin
+    #         run_thermal_bubble_golden_master_tests()
+    #     end
+    #
+    #     @testset "Deep convection" begin
+    #         run_deep_convection_golden_master_tests()
+    #     end
+    # end
 end # Oceananigans tests
